backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pandas
numpy
langchain
langchain-groq
groq
tabulate
python-dotenv
